Fix category not showing up when re-enabled
Add cascade delete rule for articles of a category. when a category is re-enabled without this, it will not be linking to any articles (since those articles are already in DB, insert/relinking to this category will not happen), thus the re-enabled category will not be visible (since we don't show a category with no articles).

Signed-off-by: Aravinth Chandran <arvinthchandran@gmail.com>