Fix crash on first launch
This crash happens when there is a change in app ID or domain

Reason: Accessing KonotorCustomProperty from a serial queue seems to be the issue but it should have also not worked properly in concurrent as well. because core data manipulations are not happening inside performblock.

More Info: https://stackoverflow.com/questions/19184262/core-data-error-when-deleting-adding-objects

Fixes: https://github.com/freshdesk/hotline_ios_dev/issues/282